split logic from UI
